requirements.txt
accelerate==1.4.0
torchdiffeq==0.2.5
torchaudio==2.6.0
x-transformers==2.1.2
transformers==4.49.0
librosa==0.10.2.post1
pyarrow==19.0.1
pandas==2.2.3
pylance==0.23.2
ema-pytorch==0.7.7
prefigure==0.0.10
bitsandbytes
muq==0.1.0
mutagen==1.47.0
pykakasi==2.3.0
jieba==0.42.1
cn2an==0.5.23
pypinyin==0.53.0
onnxruntime
Unidecode==1.3.8
phonemizer==3.3.0
inflect==7.5.0

fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic>=2.0.0
py3langid==0.2.2


